requirements.txt
opencv-python
onnxruntime-gpu
pyqt5
PyQt5Designer
pyqt5-tools
configobj


